No more roadkill
Finally managed to work out why we'd occassionally get a chunk taken
out of the middle of the lovely curves, turns out it's because we
were double chomping the central pixel in some instances.

Now breaks out of the alpha setting loop if it detects it's about to
eat the same pixel. Exiting the loop is acceptable as the only way
the condition is true is if all other pixels on either side (of the
center) have already been processed.